pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "bsedata>=0.6.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "keras>=2.14.0",
    "nltk>=3.9.1",
    "nsepython>=2.94",
    "nsetools>=2.0.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
    "websockets>=15.0.1",
    "yfinance>=0.2.55",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "openai>=1.68.2",
    "pandas-ta>=0.3.14b0",
    "ta-lib-easy>=1.0.3",
]
